@@ -6,15 +6,20 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "antd": "^5.24.6",
     "axios": "^0.21.4",
     "bootstrap": "^4.6.2",
+    "date-fns": "^4.1.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
+    "moment": "^2.30.1",
+    "react": "^19.1.0",
     "react-bootstrap": "^1.6.8",
-    "react-dom": "^17.0.2",
+    "react-datepicker": "^8.2.1",
+    "react-dom": "^19.1.0",
+    "react-icons": "^5.5.0",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-router-dom": "^5.3.4",
-    "react-scripts": "^4.0.3",
+    "react-router-dom": "^7.5.0",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
@@ -43,6 +48,8 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@babel/preset-env": "^7.26.9",
     "msw": "^0.35.0"
   },
   "engines": {
